Sistemate tutte le classi che fanno riferimento ad un entità (animate e movibili) e sistemato pippo

+++ Progetto-Tesi/core/src/com/game/graphics/entities/AnimatedEntity.java
@@ -38,7 +38,7 @@ public class AnimatedEntity extends Entity
 	{
 		sprite.setTexture(textureAtlas.findRegion(textureRegionName).getTexture());
 		textureRegions.put(textureRegionName, new TextureRegion(sprite.getTexture(),
-				0, 0, textureDimension.x, textureDimension.y));
+				0, 0, (int)textureDimension.x, (int)textureDimension.y));
 		sprite.setBounds(0, 0, textureDimension.x / AdventureGame.pixelPerMeter,
 				textureDimension.y / AdventureGame.pixelPerMeter);
 		sprite.setRegion(textureRegions.get(textureRegionName));
